pyproject.toml
[tool.poetry]
name = "sandbox"
version = "0.1.0"
description = ""
authors = ["WolfEYc <wolfeycode@gmail.com>"]
readme = "README.md"
packages = [{ include = "api" }]

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.31.0"
seaborn = "^0.13.0"
locust = "^2.19.0"
python-dotenv = "^1.0.0"
cachetools = "^5.3.2"
polars = "^0.20.2"
adbc-driver-postgresql = "^0.8.0"
pyarrow = "^14.0.2"
fastapi = "^0.105.0"
uvicorn = "^0.25.0"
asyncpg = "^0.29.0"

[tool.poetry.scripts]
main = "api.app:main"

[tool.poetry.group.dev.dependencies]
ruff = "^0.1.6"
ipykernel = "^6.27.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
